Create Class.cpp
That is a funny solution. This problem's name is class.  All of the hacker is use class to solve this problem.  But, I build my  logic without any class.  And my solution is all test cases passed.  And the runtime of mine is very small.  I guess, This is a very easy and good solution  from the all sides.